requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub